# make update_row_maps! type stable