I am slowly moving towards #2340. For now a small fix in the "core code", that shaves off some allocations because it makes it type stable and in consequence time. Nothing big, but not problematic to add either (and it should not degrade performance in any scenario).

Benchmarks:

Code setup:
```
using DataFrames
using Random
Random.seed!(1234);

df1 = DataFrame(id1 = rand(1:1000, 1000000), id2 = rand(1:1000, 1000000), x1=1:1000000);
df2 = DataFrame(id1 = rand(1:1000, 1000000), id2 = rand(1:1000, 1000000), x2=1:1000000);
```

Before PR:
```
julia> @benchmark innerjoin($df1, $df2, on=[:id1, :id2])
BenchmarkTools.Trial: 
  memory estimate:  299.87 MiB
  allocs estimate:  3999662
  --------------
  minimum time:     769.417 ms (2.13% GC)
  median time:      815.853 ms (4.77% GC)
  mean time:        830.811 ms (6.42% GC)
  maximum time:     878.898 ms (10.63% GC)
  --------------
  samples:          7
  evals/sample:     1

julia> @benchmark outerjoin($df1, $df2, on=[:id1, :id2])
BenchmarkTools.Trial: 
  memory estimate:  428.35 MiB
  allocs estimate:  3999693
  --------------
  minimum time:     1.083 s (3.58% GC)
  median time:      1.147 s (6.97% GC)
  mean time:        1.154 s (8.42% GC)
  maximum time:     1.233 s (14.65% GC)
  --------------
  samples:          5
  evals/sample:     1
```

After PR:
```
julia> @benchmark innerjoin($df1, $df2, on=[:id1, :id2])
BenchmarkTools.Trial: 
  memory estimate:  132.03 MiB
  allocs estimate:  182
  --------------
  minimum time:     691.800 ms (0.17% GC)
  median time:      715.135 ms (0.46% GC)
  mean time:        719.786 ms (1.93% GC)
  maximum time:     775.350 ms (10.40% GC)
  --------------
  samples:          8
  evals/sample:     1

julia> @benchmark outerjoin($df1, $df2, on=[:id1, :id2])
BenchmarkTools.Trial: 
  memory estimate:  260.51 MiB
  allocs estimate:  213
  --------------
  minimum time:     977.176 ms (0.06% GC)
  median time:      990.954 ms (0.87% GC)
  mean time:        997.370 ms (1.81% GC)
  maximum time:     1.043 s (6.63% GC)
  --------------
  samples:          6
  evals/sample:     1
```

So it is ~10% speedup.