@bkamins 2020-08-17
@nalimilan - this is a minimal change, but I would merge it as it is an easy fix so that we have a bit cleaner code to start major improvements.